Added notes and api client
Added some notes to "first attempt" to check if our angular framework works with JSON nested types (and it does, at least for a basic sample).

Also added an API client which I'll flesh out to show a pure pass-through front-end, and afterwards we'll explore a more advanced "internal API".